refactor: extract listCreate as a SFC
